Implement basic beacon touching rewards
Every touch is a gamble with a variety of mechanics from
obtaining the Dawnbreaker to being locked out of touching
the beacon by losing it. Electrum rewards are given for this
to incentivize; electrum will be used for a later function.
Data models updated to include data regarding per-user beacon touch
tracking.